profile images show in edit activity
